[FIX] fixes win32 compile issues

+++ src/openms/include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/EmgFitter1D.h
@@ -91,7 +91,10 @@ protected:
     {
       public:
         EgmFitterFunctor(int dimensions, const EmgFitter1D::Data * data)
-      : LevMarqFitter1D::GenericFunctor(dimensions, data->n), m_data(data) {}
+          : LevMarqFitter1D::GenericFunctor(dimensions,
+                                            static_cast<int>(data->n)),
+            m_data(data)
+        {}
 
       int operator()(const Eigen::VectorXd &x, Eigen::VectorXd &fvec);
       // compute Jacobian matrix for the different parameters
